+++ src/lossless_hermes/engine/__init__.py
@@ -190,7 +190,7 @@ class ContextEngineInfo:
     """
 
     name: str = "lcm"
-    version: str = "0.1.1"
+    version: str = "0.1.2"
     owns_compaction: bool = True